Functional but not yet Optimal game

+++ Astral Autem/Assets/Astral Autem/Scripts/SpawnerData.cs
@@ -3,7 +3,7 @@
 namespace FG {
 	[CreateAssetMenu(fileName = "SpawnerData", menuName = "ScriptableObject/EnemySpawner")]
 	public class SpawnerData : ScriptableObject {
-		public float _timeUntilWaveStart;
+		public int _timeUntilWaveStart;
 		public float _waitTimeBetweenSpawns;
 		public GameObject _enemy;
 		public float _enemyCount;
